+++ src/data_assets/assets/github/repos.py
@@ -16,7 +16,7 @@ from data_assets.assets.github.helpers import GitHubOrgAsset
 from data_assets.core.column import Column, Index
 from data_assets.core.registry import register
 from data_assets.core.types import PaginationState
-from sqlalchemy import Boolean, DateTime, Integer, Text
+from sqlalchemy import BigInteger, Boolean, DateTime, Text
 
 
 @register
@@ -33,7 +33,7 @@ class GitHubRepos(GitHubOrgAsset):
     org_request_params = {"type": "all"}
 
     columns = [
-        Column("id", Integer(), nullable=False),
+        Column("id", BigInteger(), nullable=False),
         Column("full_name", Text(), nullable=False),
         Column("name", Text()),
         Column("owner_login", Text()),
